Keep chain of notifying objects for the _flush call.
The _flush method is used to process notification from some underlying
parameter and takes care of invalidating dependent value.  It was called
with an argument of the last notifying parameter object, but that does
not show what changed for deeper parameter hierarchies.

Now the _flush argument is a tuple of the notifying parameters, the
most recent first.  Generator may now override the _flush method and
ignore any parameter changes that do not affect the generated profile.