Convert ParameterProxy to Parameter subclass.
Use explicit wrapper methods for functions that need to be forwarded
to the proxied parameters.  Remove unnecessary overrides of the
Parameter methods.  Remove __getattr__ sorcery, which made visitor
pattern behave as if handling the proxied parameter and prevented
update of profile variables in the residual equation after setProfile.

This resolves #17.